requirements.txt
streamlit>=1.32
pandas>=2.1
numpy>=1.26
plotly>=5.20
scipy>=1.11
requests>=2.31
python-dotenv>=1.0
